Added function that removes all the user's tasks in a particular garden when they are removed from that garden

+++ src/main/java/com/example/gardenplanner/controller/GardenManagementController.java
@@ -354,8 +354,12 @@ public class GardenManagementController {
      */
     private void removeUser(IPerson person)
     {
-        gardenUsersDAO.removePersonFromGarden(person, garden);
-        syncPeople();
+        if (displayConfirmPopup("Are you sure you want to remove this person from your garden?"))
+        {
+            gardenUsersDAO.removePersonFromGarden(person, garden);
+            taskDAO.deleteUserTasks(person, garden);
+            syncPeople();
+        }
     }
 
     /**
